fix : 게스트 유저 자격 추가

+++ src/main/java/efub/toy2/papers/domain/folder/controller/FolderController.java
@@ -5,7 +5,10 @@ import efub.toy2.papers.domain.folder.dto.FolderRequestDto;
 import efub.toy2.papers.domain.folder.dto.FolderResponseDto;
 import efub.toy2.papers.domain.folder.service.FolderService;
 import efub.toy2.papers.domain.member.domain.Member;
+import efub.toy2.papers.domain.member.service.MemberService;
 import efub.toy2.papers.global.config.AuthUser;
+import efub.toy2.papers.global.exception.CustomException;
+import efub.toy2.papers.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -18,16 +21,19 @@ import java.util.List;
 @RequestMapping("/folders")
 public class FolderController {
     private final FolderService folderService;
+    private final MemberService memberService;
 
     /* 폴더 생성 */
     @PostMapping
     public FolderResponseDto createFolder(@AuthUser Member member , @RequestBody FolderRequestDto requestDto){
+        if(!memberService.isAdminMember(member)) throw new CustomException(ErrorCode.NON_LOGIN);
         return folderService.createFolder(member,requestDto);
     }
 
     /* 폴더 삭제 */
     @DeleteMapping("/{folderId}")
     public String deleteFolder(@AuthUser Member member, @PathVariable Long folderId){
+        if(!memberService.isAdminMember(member)) throw new CustomException(ErrorCode.NON_LOGIN);
         return folderService.deleteFolder(member,folderId);
     }
 
@@ -41,6 +47,7 @@ public class FolderController {
     @PutMapping("/{folderId}")
     public FolderResponseDto updateFolderName(@AuthUser Member member ,
                                               @PathVariable Long folderId, @RequestBody FolderRequestDto requestDto){
+        if(!memberService.isAdminMember(member)) throw new CustomException(ErrorCode.NON_LOGIN);
         return folderService.updateFolderName(member,folderId , requestDto);
     }
 }
